+++ src/consumer/exceptions/driver-not-found.js
@@ -1,0 +1,10 @@
+export default class DriverNotFoundError extends Error {
+  driver: string;
+  lang: string;
+
+  constructor(driver: string, lang: string) {
+    super();
+    this.driver = driver;
+    this.lang = lang;
+  }
+}